Disable form filters in html. Because disabling via javascript works late.
