+++ js/util.js
@@ -59,6 +59,14 @@ const getMinPrice = (accommodationType) => {
   }
 };
 
+const OFFER_TYPES = {
+  'flat': 'Квартира',
+  'bungalow': 'Бунгало',
+  'house': 'Дом',
+  'palace': 'Дворец',
+  'hotel': 'Отель'
+};
+
 const padZero = (num) => num.toString().length <= 1 ? `0${num}` : num;
 
-export {getRandomInt, getRandom, getRandomElement, getRandomString, padZero, getMinPrice};
+export {getRandomInt, getRandom, getRandomElement, getRandomString, padZero, getMinPrice, OFFER_TYPES};
